feat: Leaderboard with Redis
- Use redisZ as leaderboard values
- Create new implementation for GetLeaderboard
- Remove streak LeaderboardEntry as it is unnecessary
- Remove AddPoints & GetLeaderboard from progress postgres repo
- Replace postgres implementation in leaderboard_repo file with a redis
implementation
- .Err() ignores the addedCount and only returns the error, if there is
any
- Implement GetPlayers to use pagination. The offset is zero-indexed,
and it is the current user's number we're at. In page 1, the Start is 0
and the Stop is 0 + 10 - 1 (if the limit is 10). And for page 2, the
Start is 10, and the Stop is 19. And so on until the leaderboard ends.

+++ src/internal/progress/handlers/http/handlers.go
@@ -61,7 +61,7 @@ func (h *ProgressHandler) GetLeaderboardHandler(echoCtx *echo.Context) error {
 		limit = 25
 	}
 
-	leaderboard, err := h.progressService.GetLeaderboard(echoCtx.Request().Context(), offset, limit)
+	leaderboard, err := h.progressService.GetLeaderboard(echoCtx.Request().Context(), int64(offset), int64(limit))
 	if err != nil {
 		return echoCtx.JSON(http.StatusInternalServerError, map[string]string{"error": "Failed to load leaderboard"})
 	}
